# swagger.yaml
---
swagger: "2.0"
info:
  title:
    Fn::Sub: scheduler-${ENVIRONMENT}
  description: REST APIs for scheduler service
  contact:
    name: Joakim Hedlund
  version: 1.0.0
servers:
  - url:
      Fn::Sub: https://aws.triplehead.net/${PROJECT}/
    description: The best of servers.
schemes:
  - https

x-definitions:
  APILambda: &APILambda
    x-amazon-apigateway-integration:
      uri:
        Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ApiLambda.Arn}/invocations
      passthroughBehavior: when_no_match
      httpMethod: POST
      type: aws_proxy

  CORSHeaders: &CORSHeaders
    Access-Control-Allow-Headers:
      type: string
    Access-Control-Allow-Methods:
      type: string
    Access-Control-Allow-Origin:
      type: string

  enableCORS: &enableCORS
    <<: *APILambda
    responses:
      200:
        description: Allows CORS

definitions:
  Job:
    type: object
    properties:
      id:
        type: string
        description: UUID for the job, e.g. `193a3b0d-aec5-42ae-acf9-40d9e843447c`
      endpoint:
        type: string
        description: URL that will be called when the job is executed.
      # TODO: add "method" and "body"?
      scheduleAt:
        type: string
        description: An ISO 8601-compatible timestamp, e.g. `2021-09-03T18:45:00.000Z`
      createdAt:
        type: string
      updatedAt:
        type: string
      executedAt:
        type: string

paths:
  /jobs:
    options:
      <<: *enableCORS
    get:
      <<: *APILambda
      summary: List all jobs
      responses:
        200:
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/Job'

    post:
      <<: *APILambda
      summary: Create a job
      consumes:
        - application/json
      parameters:
        - name: event
          in: body
          required: true
          schema:
            $ref: '#/definitions/Job'
            required:
              - endpoint
              - scheduleAt
      responses:
        200:
          description: maximum swag
        400:
          description: no bueno
  /jobs/{id}:
    options:
      <<: *enableCORS
      parameters:
        - name: id
          in: path
          type: string
          required: true
    get:
      <<: *APILambda
      summary: Retrieve a job
      parameters:
        - name: id
          in: path
          type: string
          required: true
      responses:
        200:
          description: Success
          headers:
            <<: *CORSHeaders
          schema:
            $ref: '#/definitions/Job'


x-amazon-apigateway-request-validator: validate-body-and-params
x-amazon-apigateway-request-validators:
  validate-body-and-params:
    validateRequestParameters: true
    validateRequestBody: true
x-amazon-apigateway-gateway-responses:
  DEFAULT_4XX:  &gatewayCorsResponse
    responseParameters:
      gatewayresponse.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
      gatewayresponse.header.Access-Control-Allow-Methods: "'OPTIONS,GET,POST'"
      gatewayresponse.header.Access-Control-Allow-Origin: "'*'"
  DEFAULT_5XX: *gatewayCorsResponse